+++ Fabric/src/main/java/fuzs/puzzleslib/mixin/client/MinecraftFabricMixin.java
@@ -1,5 +1,6 @@
 package fuzs.puzzleslib.mixin.client;
 
+import fuzs.puzzleslib.api.client.event.v1.BuildCreativeContentsCallback;
 import fuzs.puzzleslib.api.client.event.v1.FabricClientEvents;
 import fuzs.puzzleslib.api.client.event.v1.FabricScreenEvents;
 import fuzs.puzzleslib.api.event.v1.core.EventResult;
@@ -11,7 +12,11 @@ import net.minecraft.client.gui.screens.TitleScreen;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.multiplayer.MultiPlayerGameMode;
 import net.minecraft.client.player.LocalPlayer;
+import net.minecraft.core.NonNullList;
 import net.minecraft.network.Connection;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.CreativeModeTab;
+import net.minecraft.world.item.ItemStack;
 import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -82,5 +87,14 @@ public abstract class MinecraftFabricMixin {
         Connection connection = this.player.connection.getConnection();
         Objects.requireNonNull(connection, "connection is null");
         FabricClientEvents.PLAYER_LOGGED_OUT.invoker().onLoggedOut(this.player, this.gameMode, connection);
+    }
+
+    @ModifyVariable(method = "createSearchTrees", at = @At(value = "INVOKE", target = "Lnet/minecraft/core/NonNullList;forEach(Ljava/util/function/Consumer;)V", shift = At.Shift.BEFORE))
+    public NonNullList<ItemStack> createSearchTrees(NonNullList<ItemStack> items) {
+        for (CreativeModeTab tab : CreativeModeTab.TABS) {
+            ResourceLocation identifier = BuildCreativeContentsCallback.tryCreateIdentifier(tab);
+            FabricClientEvents.BUILD_CREATIVE_CONTENTS.invoker().onBuildCreativeContents(identifier, tab, items::add);
+        }
+        return items;
     }
 }